creation of multible geometries dependent on ispc byte limit

+++ plugins/OSPRay_plugin/include/OSPRay_plugin/CallOSPRayStructure.h
@@ -84,7 +84,7 @@ public:
     unsigned int vertexStride;
     unsigned int colorLength;
     unsigned int colorStride;
-    unsigned int partCount;
+    long long int partCount;
     float globalRadius;
     core::moldyn::SimpleSphericalParticles::ColourDataType mmpldColor;
